Determine anaemia status for various population groups

This set of functions identifies anaemia in both individual-specific target groups and the overall population, which includes all target groups' functions.

find_anaemia_u5(x)

find_anaemia_5to11(x)

find_anaemia_12to14(x)

find_anaemia_np_women(x)

find_anaemia_pregnant(x)

find_anaemia_men(x)

find_anaemia(
  df,
  group = c("u5", "5to11", "12to14", "np_women", "pregnant", "men"),
  hb = NULL,
  add = TRUE
)

Arguments

x

A numeric vector containing haemoglobin values in grams per litre (g/L) unit format for a specific population group. The following table provides the specific command based on the respective population group.

Population	Command

Children 6-59 months of age	find_anaemia_u5

Children 5-11 years of age	find_anaemia_5to11

Children 12-14 years of age	find_anaemia_11to14

Non-pregnant women - 15 years and above	find_anaemia_np_women

Pregnant women	find_anaemia_pregnant

Men - 15 years and above	find_anaemia_men

df

A data.frame of survey dataset with information on haemoglobin value and information on population group to which the sample comes from.

group

A character value specifying the population target group to identify anaemia from. The function can determine the anaemia status for the following groups:

Population	Short-form

Children 6-59 months of age	"u5"

Children 5-11 years of age	"5to11"

Children 12-14 years of age	"11to14"

Non-pregnant women - 15 years and above	"np_women"

Pregnant women	"pregnant"

Men - 15 years and above	"men"

The short-form is used in the function to indicate which population group to find anaemia from.

hb

A characater value specifying the variable name in df containing the sample observation's haemoglobin level. Note that the haemoglobin values should be recorded in grams per liter (g/L) units. If NULL (default), will attempt to find variable in df that is likely to hold the haemoglobin levels information by searching for most common variable names used for haemoglobin. If none can be found, will issue an error that appropriate haemoglobin values cannot be found.

add

Logical. Should anaemia classification results be added to df? Default to TRUE which will add a variable named anaemia_status to df.

Value

If add is TRUE, a data frame with the same structure as df and with a new variable anaemia_status otherise a vector of class factor. The new variable or the resulting vector provides anaemia categories to each observation: no anaemia, mild anaemia, moderate anaemia, and severe anaemia. The following table explains the cut-off points to classify each observation accordingly:

Population	Mild	Moderate	Severe

Children 6-59 months of age	100 - 109	70 - 99	< 70

Children 5-11 years of age	110 - 114	80 - 109	< 80

Children 12-14 years of age	110 - 119	80 - 109	< 80

Non-pregnant women

15 years and above	110 - 119	80 - 109	< 80

Pregnant women	100 - 109	70 - 99	< 70

Men

15 years and above	110 - 129	80 - 109	< 80

Author

Nicholus Tint Zaw

Examples

 # Create testing data
 x <- runif(20, min = 60, max = 130)

 hb <- runif(50, min = 60, max = 130)
 gender <- rep(c("male", "female"), each = 25)

 df <- data.frame(gender, hb)

 # For individual target group function;
  find_anaemia_u5(x)          # U5 Children
#>  [1] no anaemia       moderate anaemia no anaemia       no anaemia      
#>  [5] severe anaemia   moderate anaemia moderate anaemia moderate anaemia
#>  [9] moderate anaemia moderate anaemia moderate anaemia severe anaemia  
#> [13] severe anaemia   moderate anaemia no anaemia       moderate anaemia
#> [17] no anaemia       no anaemia       no anaemia       mild anaemia    
#> Levels: severe anaemia moderate anaemia mild anaemia no anaemia
  find_anaemia_5to11(x)       # Children 5 - 11 years
#>  [1] no anaemia       moderate anaemia no anaemia       no anaemia      
#>  [5] severe anaemia   moderate anaemia severe anaemia   moderate anaemia
#>  [9] moderate anaemia moderate anaemia moderate anaemia severe anaemia  
#> [13] severe anaemia   severe anaemia   mild anaemia     moderate anaemia
#> [17] no anaemia       no anaemia       mild anaemia     moderate anaemia
#> Levels: severe anaemia moderate anaemia mild anaemia no anaemia
  find_anaemia_12to14(x)      # Children 12 - 14 years
#>  [1] no anaemia       moderate anaemia mild anaemia     no anaemia      
#>  [5] severe anaemia   moderate anaemia severe anaemia   moderate anaemia
#>  [9] moderate anaemia moderate anaemia moderate anaemia severe anaemia  
#> [13] severe anaemia   severe anaemia   mild anaemia     moderate anaemia
#> [17] mild anaemia     mild anaemia     mild anaemia     moderate anaemia
#> Levels: severe anaemia moderate anaemia mild anaemia no anaemia
  find_anaemia_np_women(x)    # Non-pregnant Women
#>  [1] no anaemia       moderate anaemia mild anaemia     no anaemia      
#>  [5] severe anaemia   moderate anaemia severe anaemia   moderate anaemia
#>  [9] moderate anaemia moderate anaemia moderate anaemia severe anaemia  
#> [13] severe anaemia   severe anaemia   mild anaemia     moderate anaemia
#> [17] mild anaemia     mild anaemia     mild anaemia     moderate anaemia
#> Levels: severe anaemia moderate anaemia mild anaemia no anaemia
  find_anaemia_pregnant(x)    # Pregnant Women
#>  [1] no anaemia       moderate anaemia no anaemia       no anaemia      
#>  [5] severe anaemia   moderate anaemia moderate anaemia moderate anaemia
#>  [9] moderate anaemia moderate anaemia moderate anaemia severe anaemia  
#> [13] severe anaemia   moderate anaemia no anaemia       moderate anaemia
#> [17] no anaemia       no anaemia       no anaemia       mild anaemia    
#> Levels: severe anaemia moderate anaemia mild anaemia no anaemia
  find_anaemia_men(x)         # Men
#>  [1] mild anaemia     moderate anaemia mild anaemia     mild anaemia    
#>  [5] severe anaemia   moderate anaemia severe anaemia   moderate anaemia
#>  [9] moderate anaemia moderate anaemia moderate anaemia severe anaemia  
#> [13] severe anaemia   severe anaemia   mild anaemia     moderate anaemia
#> [17] mild anaemia     mild anaemia     mild anaemia     moderate anaemia
#> Levels: severe anaemia moderate anaemia mild anaemia no anaemia

 # For overall population function;
  find_anaemia(df = df,
               hb = "hb",
               add = TRUE)
#>    gender        hb   anaemia_status
#> 1    male 114.89823       no anaemia
#> 2    male  85.92006 moderate anaemia
#> 3    male 106.43402     mild anaemia
#> 4    male  81.88556 moderate anaemia
#> 5    male  89.43919 moderate anaemia
#> 6    male 124.00305       no anaemia
#> 7    male 103.75630     mild anaemia
#> 8    male 100.75331     mild anaemia
#> 9    male 128.87569       no anaemia
#> 10   male  73.18953 moderate anaemia
#> 11   male 108.53991     mild anaemia
#> 12   male 101.22482     mild anaemia
#> 13   male  88.13181 moderate anaemia
#> 14   male 105.22536     mild anaemia
#> 15   male  60.40579   severe anaemia
#> 16   male  75.58501 moderate anaemia
#> 17   male 107.63446     mild anaemia
#> 18   male 108.83131     mild anaemia
#> 19   male 111.03913       no anaemia
#> 20   male  79.45960 moderate anaemia
#> 21   male  82.36316 moderate anaemia
#> 22   male 127.86357       no anaemia
#> 23   male  89.55571 moderate anaemia
#> 24   male  95.31509 moderate anaemia
#> 25   male  97.56395 moderate anaemia
#> 26 female  68.69837   severe anaemia
#> 27 female  85.42423 moderate anaemia
#> 28 female  70.36817 moderate anaemia
#> 29 female  95.46847 moderate anaemia
#> 30 female 105.76180     mild anaemia
#> 31 female  65.25176   severe anaemia
#> 32 female 105.24539     mild anaemia
#> 33 female  70.05021 moderate anaemia
#> 34 female  90.92743 moderate anaemia
#> 35 female  84.48434 moderate anaemia
#> 36 female 106.20693     mild anaemia
#> 37 female 120.19861       no anaemia
#> 38 female  64.97770   severe anaemia
#> 39 female 110.76267       no anaemia
#> 40 female  95.08430 moderate anaemia
#> 41 female 119.00001       no anaemia
#> 42 female 128.41554       no anaemia
#> 43 female  79.07634 moderate anaemia
#> 44 female  88.01670 moderate anaemia
#> 45 female  65.81287   severe anaemia
#> 46 female 113.33153       no anaemia
#> 47 female 117.50616       no anaemia
#> 48 female 125.52413       no anaemia
#> 49 female  76.95117 moderate anaemia
#> 50 female  66.62928   severe anaemia