Anaemia Diagnosis

Perform anaemia diagnosis based on age, sex and pregnancy status

flag_anaemia(
  df,
  age = NULL,
  hb = NULL,
  sex = NULL,
  pregnant = NULL,
  add = TRUE
)

Arguments

df

Survey dataset (as an R data.frame) with the following variables present; sex, age, haemoglobin value, and pregnancy status. If the study sample population did not contain the pregnancy population, the pregnancy status (dummy variable yes/no) might not be required for this function. But, the other variables are compulsory for function processing. If the dataset did not contain those variables, please perform data processing to have those variables in the dataset.

age

Sample observation's age from the dataset. This variable should be recorded in the year format. If your dataset has only an age variable with month, please converts it into year format before using this function.

hb

Sample observation's haemoglobin level from the dataset. The Hb values should record in the grams per liter (g/L) units.

sex

Sample observation's sex identifies from the dataset. This variable should be coded in the dummy variable as 1 for "male" and 0 for "female". If your dataset's respondent sex variable was not recorded this way, please make necessary data recoding before using this function.

pregnant

Sample observation's pregnancy status from the dataset. This variable should be coded as a dummy variable: 1 for pregnant and 0 for non-pregnant. Please perform the necessary data processing for recoding if your dataset observation's pregnancy status was not in dummy variable format.

add

This parameter's default option is TRUE and will add new generated variables anaemia_all to your existing dataset applied in this function. This newly developed categorical variable comprises three types of flag categories resulting from data checking results; no anaemia, mild anaemia, moderate anaemia, and severe anaemia. The following table explains the cut-off points applied in this diagnostic function.

Population	Mild	Moderate	Severe

Children 6-59 months of age	100 - 109	70 - 99	< 70

Children 5-11 years of age	110 - 114	80 - 109	< 80

Children 12-14 years of age	110 - 119	80 - 109	< 80

Non-pregnant women

(15 years and above)	110 - 119	80 - 109	< 80

Pregnant women	100 - 109	70 - 99	< 70

Men

(15 years and above)	110 - 129	80 - 109	< 80

Value

A data frame with the same structure as df. The new variable anaemia_all can be observed, containing the information the type of anaemia present: mild, moderate, or severe.